Change player struct to reflect heading of vector.
This is preferable to simply displaying the rotation, which was altered by pushing left and right. The heading is calculated from the player's location and an arbitrary point ahead of it, using velocity. I think this might not be the best way of representing the player. After all, if we're re-calculating a 'new location' by applying the velocity to the location, then we need to alter the tick function to update this heading value. It might be worth looking into getting rid of the heading entirely and changing the velocity directly by adding/subtracting another vector containing cos(R), sin(R), where R := Turn Rate, and treating the "speed" as the magnitude or radius of the vector or circle surrounding the ship. Then we only pull the heading from the vector when we need to render the ship, and the velocity reflects the ship's direction.