Changing rotation to use an indexed unit circle.
Instead of adding cos(π/6), sin(π/6) to the velocity, we simply index a unit circle. 0 is the base, facing right. 3 is 90̣° or π/2. Pressing "←" simply adds 1 to the index, pressing "→" subtracts one. The result is modulo M, where "M" is the maximum amount of rotations around a unit circle for the given turn-rate. When the turn-rate is π/6, there are 12 rotations possible before rolling back over. The default velocity vector has a magnitude of 1 and represents the unit circle. Every clock tick, the radius is multiplied by the current speed, scaling the vector to the appropriate speed while keeping rotation simple.