Finish index based rotation.
Now the Player struct tracks the amount of rotations made from 0 and uses this information for rendering and movement. Pressing left rotates the ship's index counter-clockwise, and pressing right rotates it clockwise. There's still some tweaking left to do. Right now, racket interprets rotation where "0" means no rotation, but the turning-tracking interprets "0" to mean pointed right, like the trigonometric unit circle. In addition, Racket treats the upper-left as the origin, which transforms the quadrants. Where the upper-right is +,+ in trig, in Racket it's +,- since you're moving upwards (negative movement) and to the right (positive). Just changing the y-axis is enough, in both Racket and trig moving right is positive. The game, when run, also begins in a separate thread. This allows you to continue using the REPL when the game is running.

At some point, I should tweak the index-based structure or how it's interpreted. Right now 90 degrees or π/2 is hardcoded into the game.