Cleanup, add insertion/retrieval functions.
Nodes now include content, but are empty lists by default. Added some comments about using vectors instead of lists in order to take advantage of indices. The 0 index could be the upper-left, the 1 index could be the upper right, etc. until the 3 index is the lower right. I simplified make-subtrees. I didn't need to use folding, just a map function.

Zero depth is now corrected and represents the root node.

Added functions for retrieving and inserting nodes. The functions are first-come first-serve and need to be fixed, but this is just the prototype. They need to be fixed to insert and retrieve to appropriate nodes, probably by taking dimension or index into account. This is a result of the ambiguity resulting from allowing the upper-left point to represent the entire quad. For example, the point (0, 0) represents every node starting at that point including the largest (the entire screen/root-node) and the smallest (an infinitesimally small quad).