Refactor sub-trees, add comments.
Make-Sub-Trees was refactored into a single filter-map function, this way we can filter the list as we create it instead of having to create the list of sub-trees, then re-traverse it to filter out empty trees.

Add in some ideas about alternative quadtree structuring. We could create the entire tree at once, and then pass it around to insert/retrieve as needed. I think it could be better to just load the root node first with the appropriate dimension, then create by splitting the node into subnodes when needed, then inserting an entity.

Added note about retrieve-node, dimension/bounds relates to depth: Where "B" is the max bounds the quad was generated with, and "n" is the depth, dimension at any level is B^1/n. That might be useful if I use the 2nd data structure, where the whole tree is loaded at once to max depth we need to use the dimension to determine which depth to look at to find the appropriate node. The same knowledge can be used during insertion. However, if we split the node when needed instead of all at once, then we can use an insertion/retrieval strategy where we pull the LAST node it could possibly be in (the first will always be root, representing the whole screen).