Add function for retrieving index of an entity in a quadrant.
Currently this is point-based, not area based, and favors left-side over right. If a ship is centered on the line between quads 0 and 1, it will be listed as being in the 0th quadrant. Developing a point-based quadtree and then modifying it to be area/interval based later is the plan.