Add quad-insertion function and node-splitting.
Split takes a root node, and splits it into a node with four empty subnodes.
Insert-node takes a root node, and inserts the entity into the appropriate sub-node, placing the entity into the root node if it doesn't fit in any indexed quad, and splitting the appropriate child-node if it's full.