Refactor posns, quads, add retrieval.
Operations on an arbitrary multiple amount of posns are abstracted over
with posn-oper.

Indexing now properly handles potentially empty lists of subnodes.

Likewise with insertion, it checks for a #f index and places into root
node, otherwise places into appropriate subnode.

Retrieve-node gets the index of the entity, then collects the content
of that node with any entities it may collide with.

There's still a lot of refactoring to do. I get a funny feeling
about the fact that I have to carry around the bounds all the time,
there's probably something that could be factored out there.
The indexing needs to be refactored as well, as it's currently point
based and doesn't take area into account.