Add data structure, refactor.
Now contains the entity struct, which contains the coordinates of an
entity and its dimensions.
Get-index and fits? now use in-bounds? instead of posn<, which correctly
handles dimension.
Node-related functions now treat bounds as an optional variable, using
BOUNDS constant if not supplied.
Retrieve-node now correctly returns list of possible entities that could
collide with a given entity.