Implement scrolling background.
The background is scrolled via successively cropping the background
image. As the player moves left, the background is "scrolled" right at
a rate proportional to the distance between the player and the center
of the screen.

Scrolling works by consuming the player's current position and half
the screen's dimensions. It subtracts that number from the position,
which determines the new upper-left corner of the screen on the
background, then scales that number by the difference change between
the player's position and the center of the screen. As you move left,
the background scrolls right.