Refactor big-bang expression.
Move handler functions direct-game, update-game inside the big-bang
expression so they can take advantage of using locally defined the base
NODE. First the position of the turret is initialized, then inserted
into the quadtree. This tree becomes the root for all future insertions.
This way you don't reinsert the turret (which never moves) into the tree
every tick, which takes more time. Unfortunately, it's still very slow.

Created function header for exceeds-quad? which should determine if a
given entity exceeds its quad and needs to be replaced. I don't think
this is a particularly good solution to the problem, you still need to
"walk the quad" every tick to find the entity, then check whether it
exceeds the area of its bounding quad. The former takes time
proportionate to the entities position in the tree, the latter is a
constant-time operation. Would this actually be much faster than just
recreating the tree by walking through it?