Optimization notes, use vectors for quadtree.
I renamed move to turn, to clarify what it does.
I stopped the game from calling render-ui and render-info. Render-info
was causing a huge slowdown, making the game virtually unplayable.
I'm not entirely sure why. I was benchmarking a lot of the functions
times over ranges from 340 to 100000 or so to find out which was taking
the most time, and left the benchmarking function in. I added my notes.

I swapped the major data-structure for child-nodes in the quadtree
module to vectors instead of lists. Vectors are inherently indexed,
which means they lend nicely to this problem where I need to frequently
refer to indices. I should refactor index/a though, it's still a
handwritten comprehension and could be replaced with for/vector or
something else. On the whole, I got around a 30% speedup just from
swapping from lists to vectors. Very nice.

I also started using incremental garbage collection.