Add turret tracking.
The turret now tracks the player around the map, rotating to keep them
in line-of-sight. Takes the arctangent of the difference between the
player's position and the turret's. Converts that from radians into
degrees, and then subtracts 90 from that to appropriately align it.

Given an x-distance and a y-distance on a cartesian plane, we can take
the inverse of the result of dividing y ÷ x (the tangent measurement),
which gives us the angle from the baseline.