Fix turret rotation, obstacles.
The turret is now a separate data structure extending the entity struct.
It contains the turret's rotation and firing information. The base
obstacle image is now placed over a transparent image, which is then
overlayed on the background during gameplay. I added some comments about
the problems of relative vs. absolute coordinates and obstacles. The
screen is "relative", where the upper left is always 0 and lower right
always 720, but the background image is much larger than that and
scrolls. Turret rotation is no longer a purely graphical change, now
being recorded in its struct for firing direction.

Readme reflects new thoughts on changes, and new tasks to do.