Remove scrolling, fix collision+quadtree.
Collision detection was mostly fixed, but still needs to be tweaked.
Sometimes if you're "close" to an entity, you get stuck, and sometimes
you fly over it entirely. Weird. I cleaned it up too and removed
redundant boolean checks, as well as making the function more clear
regarding what it's actually checking (is the left edge of one entity
between the left and right edge of the other entity? etc.) I also
abstracted over the previous big list of repeated functions with +-half,
which returns two values and is used in a define-values call to neatly
package together the appropriate measurements for each entity's edge.

I removed scrolling because it was too difficult to get it to play nice
with the quadtree, the logical locations of entities on the quadtree
didn't correspond to their absolute locations on the background and that
was breaking collision detection. I'll add it back in later. A good idea
might be to render everything over a transparent "screen" or layer,
which is static, and then layer that over a scrolling background instead
of trying to figure out how to render still objects over a moving
background every tick.

Added code for debugging, the quadtree is now rendered over the
background dynamically as nodes are created and destroyed.

Fixed a major problem in the quadtree module. I was checking of the
children of a node was empty. A vector can't be checked with empty,
which returns #f for anything except for any empty LIST. I fixed it to
properly use zero?+vector-length. I fixed the indexing function to use
for/first instead of manually stepping through a quad's child nodes.